Just before the exam!

+++ quiz/views.py
@@ -1,18 +1,19 @@
-import xml
+# import xml
 
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.contrib.auth.models import User
+from django.core.exceptions import PermissionDenied
 from django.forms import formset_factory, modelformset_factory
 from django.shortcuts import render, redirect
 from django.urls import reverse_lazy, reverse
-from django.views.generic import CreateView, FormView, DetailView, DeleteView
+from django.views.generic import CreateView, FormView, DetailView, DeleteView, ListView, UpdateView
 from django.views.generic.base import View
 import xml.etree.ElementTree as xml_tree
 
 from quiz.forms import QuizCreateForm, QuestionCreateForm, AnswerCreateForm, QuizAssignForm, TestQuizForm
 from quiz.models import Quiz, Question, Answer, StudentTest, SUBJECT, StudentTestAnswer
-from registrator.models import TeacherProfile, StudentProfile, School
+from registrator.models import TeacherProfile, StudentProfile, School, SchoolClass
 
 
 def render_home(request):
@@ -22,7 +23,7 @@ def render_home(request):
     return render(request, 'quiz/home.html', {'schools': schools, 'subjects': subjects})
 
 
-class StudentDetailView(DetailView):
+class StudentDetailView(DetailView, LoginRequiredMixin):
     model = User
     template_name = 'quiz/student-detail.html'
 
@@ -31,7 +32,7 @@ class StudentDetailView(DetailView):
         student_profile = StudentProfile.objects.get(user=self.request.user)
         assigned_tests = StudentTest.objects.filter(student=self.request.user)
         completed_tests = [test for test in assigned_tests if test.is_completed]
-        average_score = sum([test.score for test in completed_tests]) / len(completed_tests)
+        average_score = sum([test.score for test in completed_tests]) / len(completed_tests) if completed_tests else 0
         context['profile'] = student_profile
         context['assigned_tests_number'] = len(assigned_tests)
         context['completed_test_number'] = len(completed_tests)
@@ -39,7 +40,7 @@ class StudentDetailView(DetailView):
         return context
 
 
-class StudentTestsView(DetailView):
+class StudentTestsView(DetailView, LoginRequiredMixin):
     model = User
     template_name = 'quiz/student-tests.html'
 
@@ -50,20 +51,71 @@ class StudentTestsView(DetailView):
         return context
 
 
-class TeacherDetailView(DetailView):
+class AnsweredTestView(ListView, LoginRequiredMixin):
+    model = StudentTestAnswer
+    template_name = 'quiz/answered-test-detail.html'
+    context_object_name = 'test_answers'
+
+    def get_queryset(self):
+        test_id = self.kwargs['test_id']
+        return StudentTestAnswer.objects.filter(student_test=test_id)
+
+    def get_context_data(self,*args, **kwargs):
+        context = super().get_context_data(**kwargs)
+        student_test_id = self.kwargs['test_id']
+        quiz = StudentTest.objects.get(id=student_test_id).quiz
+        context['quiz'] = quiz
+        return context
+    # def get_context_data(self, **kwargs):
+    #     context = super().get_context_data(**kwargs)
+    #     test_id = self.kwargs['test_id']
+    #     test_answers =
+
+
+class TeacherDetailView(DetailView, LoginRequiredMixin):
     model = User
     template_name = 'quiz/teacher-detail.html'
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         teacher_profile = TeacherProfile.objects.get(user=self.request.user)
+        school_classes = [f'{sc.class_level}-{sc.class_letter}' for sc in teacher_profile.school_class.all()]
         quizzes = Quiz.objects.filter(author=self.request.user)
+        # school_classes = SchoolClass.objects
         context['profile'] = teacher_profile
         context['quizzes'] = quizzes
+        context['school_classes'] = school_classes
         return context
 
 
-class QuizDetailView(DetailView):
+@login_required
+def assigned_tests_view(request):
+    quiz_classes = {}
+    quizzes = Quiz.objects.filter(author=request.user)
+
+    for quiz in quizzes:
+        students_with_quiz = [student_test.student for student_test in StudentTest.objects.filter(quiz=quiz)]
+        student_profiles_with_quiz = StudentProfile.objects.filter(user__in=students_with_quiz)
+        school_classes_with_quiz = set([student_profile.school_class for student_profile in student_profiles_with_quiz])
+        quiz_classes[quiz] = school_classes_with_quiz
+    context = {'quizzes': quizzes, 'quiz_classes': quiz_classes}
+    return render(request, 'quiz/assigned-quizzes.html', context)
+
+
+class QuizSchoolClassView(ListView, LoginRequiredMixin):
+    model = StudentTest
+    template_name = 'quiz/quiz-class.html'
+    context_object_name = 'students_tests'
+
+    def get_queryset(self):
+        quiz_id = self.kwargs['quiz_id']
+        school_class_id = self.kwargs['class_id']
+        students_profiles_for_class = StudentProfile.objects.filter(school_class_id=school_class_id)
+        students_from_class = [student_profile.user for student_profile in students_profiles_for_class]
+        return StudentTest.objects.filter(student__in=students_from_class).filter(quiz_id=quiz_id)
+
+
+class QuizDetailView(DetailView, LoginRequiredMixin):
     model = Quiz
     template_name = 'quiz/quiz-detail.html'
     context_object_name = 'quiz'
@@ -76,7 +128,7 @@ class QuizDetailView(DetailView):
         return context
 
 
-class QuizCreateView(FormView):
+class QuizCreateView(FormView, LoginRequiredMixin):
     form_class = QuizCreateForm
     template_name = 'quiz/quiz-create.html'
     success_url = reverse_lazy('quiz-home')
@@ -133,7 +185,10 @@ def quiz_delete_view(request, pk):
 @login_required
 def quiz_edit_view(request, pk):
     quiz = Quiz.objects.get(id=pk)
+    if not quiz.author == request.user:
+        raise PermissionDenied
     if request.method == 'POST':
+        # permissions must be set
         form = QuizCreateForm(request.POST, instance=quiz)
         if form.is_valid():
             form.save(request.user)
@@ -168,9 +223,28 @@ class QuestionDetailView(DetailView):
         return context
 
 
-@login_required
+# class QuestionEditView(UpdateView):
+#     model = Question
+#     form_class = QuestionCreateForm
+#     template_name = 'quiz/question-edit.html'
+#
+#     def get_context_data(self, **kwargs):
+#         context = super().get_context_data(**kwargs)
+#         pk = self.kwargs['pk']
+#         context['quiz'] = Question.objects.get(id=pk).quiz
+#         return context
+
 def question_edit_view(request, pk):
-    pass
+    question = Question.objects.get(id=pk)
+    if request.method == 'GET':
+        form = QuestionCreateForm(instance=question)
+        return render(request, 'quiz/question-create.html', {'form': form,'question': question})
+    form = QuestionCreateForm(request.POST, instance=question)
+    if form.is_valid():
+        form.save(request.user, question.quiz.id)
+        return redirect('quiz-detail', question.quiz.id)
+
+
 
 
 @login_required
@@ -200,6 +274,8 @@ def answer_create_view(request, question_id):
 def answer_delete_view(request, pk):
     answer = Answer.objects.get(id=pk)
     question = answer.question
+    if question.author != request.user:
+        raise PermissionDenied
     if request.method == 'GET':
         return render(request, 'quiz/delete.html', {'answer': answer})
     answer.delete()
@@ -227,6 +303,8 @@ def quiz_assign_view(request, pk):
 @login_required
 def test_take_view(request, test_id):
     test = StudentTest.objects.get(id=test_id)
+    if test.student != request.user:
+        raise PermissionDenied
     quiz = test.quiz
     if request.method == 'GET':
         form = TestQuizForm(quiz)
